Refactor event bus
