+++ engine/lib/phx/script/meta/EventPayloadTable.lua
@@ -9,6 +9,9 @@ function EventPayloadTable.Create() end
 ---@return integer
 function EventPayloadTable:len() end
 
+---@return boolean
+function EventPayloadTable:isEmpty() end
+
 ---@param name string
 ---@return boolean
 function EventPayloadTable:contains(name) end
